Move `fetch_mojang_public_keys`
